refactor: get rid of cypress tasks for vnets
