--- tests/cypress/support/tasks/tenderly/index.ts
@@ -1,7 +1,0 @@
-export * from './types'
-export {
-  createVirtualTestnet,
-  type CreateVirtualTestnetOptions,
-  type CreateVirtualTestnetResponse,
-} from './vnet-create'
-export { deleteVirtualTestnets, type DeleteVirtualTestnetsOptions } from './vnet-delete'
